# Use jsoniter library for chunk encoding/decoding