Fixes #626.  Part of #624.

This is better than I managed to do with `ugorji/codecgen`, and substantially easier to build (no code-generation step).

Note I reverted #634 here because jsoniter defeats that trick unless you add the `jsoniter_sloppy` tag, and subsequent work suggests that large number of chunks with the same fingerprint is rare in real life.  Jsoniter still comes out faster even on the benchmark crafted to work well with #634.

Benchmarks (best of five runs):

Before:
```
BenchmarkEncode-2        	  200000	      5478 ns/op	    3092 B/op	      25 allocs/op
BenchmarkDecode1-2       	   10000	    134498 ns/op	  159409 B/op	     127 allocs/op
BenchmarkDecode100-2     	    1000	   1760706 ns/op	  648083 B/op	    2800 allocs/op
BenchmarkDecode10000-2   	      10	 171423872 ns/op	49515368 B/op	  270101 allocs/op
```

After:
```
BenchmarkEncode-2        	  300000	      3887 ns/op	    2981 B/op	       9 allocs/op
BenchmarkDecode1-2       	   20000	     59390 ns/op	  152358 B/op	      56 allocs/op
BenchmarkDecode100-2     	    1000	   1056652 ns/op	  615078 B/op	    5034 allocs/op
BenchmarkDecode10000-2   	      10	 109219380 ns/op	47543388 B/op	  523181 allocs/op
```